requirements.txt
customtkinter>=5.2
sounddevice>=0.4
faster-whisper>=1.0
transformers>=4.40
torch>=2.0
accelerate>=0.26
pyperclip>=1.8
numpy>=1.24
python-dotenv>=1.0
